refactor(bezier): move Bezier class from __init__.py into _bezier.py
